UOELSA-715: Näytä latausikoni ennen navigaation renderöintiä
- Näytä latausikoni kunnes käyttäjän tiedot on haettu storesta, jonka jälkeen navigaatio näytetään sen mukaan onko käyttäjä kirjautunut vai ei
- Korjaa 404-sivun asettelu ei-kirjautuneille käyttäjille
