+++ src/views/404/page-not-found-content.vue
@@ -1,20 +1,22 @@
 <template>
-  <b-container class="mt-4 mt-md-5 mb-6 ml-2 ml-sm-5 ml-md-4">
-    <b-row>
-      <b-col md="5">
-        <h1 class="text-primary mt-2 mt-md-5 mb-4">{{ $t('sivua-ei-loytynyt-otsikko') }}</h1>
-        <p class="mr-4 mb-4">{{ $t('sivua-ei-loytynyt-kuvaus') }}</p>
-        <div class="d-flex flex-wrap mb-4">
-          <elsa-button variant="primary" class="mr-3 mb-4" :to="{ name: 'etusivu' }">
-            {{ $t('siirry-etusivulle') }}
-          </elsa-button>
-        </div>
-      </b-col>
-      <b-col class="mr-4">
-        <img src="@/assets/elsa-404.svg" :alt="`404 ${$t('sivua-ei-loydy')}`" class="img-fluid" />
-      </b-col>
-    </b-row>
-  </b-container>
+  <b-row>
+    <b-col md="5">
+      <h1 class="text-primary mt-2 mt-md-5 mb-4">{{ $t('sivua-ei-loytynyt-otsikko') }}</h1>
+      <p class="mr-4 mb-4">{{ $t('sivua-ei-loytynyt-kuvaus') }}</p>
+      <div class="d-flex flex-wrap mb-4">
+        <elsa-button variant="primary" class="mr-3 mb-4" :to="{ name: 'etusivu' }">
+          {{ $t('siirry-etusivulle') }}
+        </elsa-button>
+      </div>
+    </b-col>
+    <b-col class="mr-4">
+      <img
+        src="@/assets/elsa-404.svg"
+        :alt="`404 ${$t('sivua-ei-loytynyt-otsikko')}`"
+        class="img-fluid"
+      />
+    </b-col>
+  </b-row>
 </template>
 
 <script lang="ts">
@@ -27,7 +29,7 @@
       ElsaButton
     }
   })
-  export default class PageNotFound extends Vue {}
+  export default class PageNotFoundContent extends Vue {}
 </script>
 
 <style lang="scss" scoped>
